requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
networkx>=2.6.0
xgboost>=1.5.0
lightgbm>=3.3.0
google-generativeai>=0.4.0
tqdm>=4.62.0
python-dotenv>=1.0.0